+++ tutorial/Game/InputManager.h
@@ -55,7 +55,7 @@
 		Renderer* rndr = (Renderer*)glfwGetWindowUserPointer(window);
 		Game* scn = (Game*)rndr->GetScene();
 
-		rndr->UpdatePosition((float)xpos,(float)ypos);
+		rndr->UpdatePosition(xpos, ypos);
 
 		if (rndr->CheckViewport(xpos,ypos, 0))
 		{
